more fixes for using a long for item index
git-svn-id: http://svn.xp-dev.com/svn/wxextension@2703 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/include/wx/extension/report/listview.h
@@ -128,7 +128,7 @@ private:
 
   bool m_ContentsChanged;
   bool m_ItemUpdated;
-  int m_ItemNumber;
+  long m_ItemNumber;
   const long m_MenuFlags;
   const ListType m_Type;
 
@@ -175,7 +175,7 @@ protected:
 private:
   void DeleteDoubles();
   const wxString GetFindInCaption(int id) const;
-  void ItemActivated(int item_number);
+  void ItemActivated(long item_number);
   void RunItems(const wxExTool& tool);
   wxExFrameWithHistory* m_Frame;
